Split off common code into libraries
