+++ tests/IResearch/IResearchAnalyzerFeatureTest.cpp
@@ -31,7 +31,7 @@
 #include "utils/utf8_path.hpp"
 
 #include "IResearch/IResearchTestCommon.h"
-#include "IResearch/RestHandlerMock.h"
+#include "Mocks/RestHandlerMock.h"
 #include "IResearch/common.h"
 #include "Mocks/LogLevels.h"
 #include "Mocks/Servers.h"
@@ -53,8 +53,8 @@
 #include "FeaturePhases/DatabaseFeaturePhase.h"
 #include "FeaturePhases/V8FeaturePhase.h"
 #include "GeneralServer/AuthenticationFeature.h"
-#include "IResearch/AgencyMock.h"
-#include "IResearch/ExpressionContextMock.h"
+#include "Mocks/AgencyMock.h"
+#include "Mocks/ExpressionContextMock.h"
 #include "IResearch/IResearchAnalyzerFeature.h"
 #include "IResearch/IResearchCommon.h"
 #include "IResearch/IResearchFeature.h"
