+++ tests/RestHandler/RestViewHandlerTest.cpp
@@ -28,7 +28,7 @@
 #include "velocypack/Builder.h"
 #include "velocypack/Parser.h"
 
-#include "IResearch/RestHandlerMock.h"
+#include "Mocks/RestHandlerMock.h"
 #include "IResearch/common.h"
 #include "Mocks/LogLevels.h"
 #include "Mocks/Servers.h"
