+++ tests/StorageEngine/PhysicalCollectionTest.cpp
@@ -27,7 +27,7 @@
 #include <velocypack/Iterator.h>
 #include <velocypack/Parser.h>
 
-#include "IResearch/RestHandlerMock.h"
+#include "Mocks/RestHandlerMock.h"
 #include "IResearch/common.h"
 #include "Mocks/LogLevels.h"
 #include "Mocks/StorageEngineMock.h"
